feat: editor with line numbers and breakpoints
Create a text editor component with line numbers on the left and a
toggleable highlight on each line number to indicate whether or not
a breakpoint is set on the line. The editor follows the cursor by
scrolling it into view if the right or lower border is reached.
The height of the line numbers match the height of the text editor
lines and follow scrolling.