feat: implement compiler service
The compiler service is responsible for compiling the source
files to bytecode, the VM understands. It caches the compilation
result. It provides all time the latest compilation result
of the highest code revision. If a result of a higher revision
is requested than the revision, stored in cache, a new compilation
process will be started. It also provides the latest succesful
compilation result.